fix: 🎨 xmake question And question modal about build pipeline in setup OSC (#506)
* fix: :art: xmake?

* fix: :ambulance: OSC setup question and 'Are you using xmake?'

During OSC setup the question about Build Pipeline only comes when there is no Build pipeline. And the question 'Are you using xmake?' only shown when they are using xmake, and it comes with the option 'yes' preselected.

* refactor: :art: Remove console.log

* refactor: :zap: Function Name change and Wrapper funtion for combining pipelinewithNotManagedServices.

Made the name of the function more clear 'moveToNextStep'. created a function 'combinePipelineWithNotManagedService' in 'pipeline.service.ts'